Move the Task in flash.py directly into pl_wrapper.py
